refactored

+++ codility/lessons/1 - iterations/binary gap/oo/binary_gap_oo.cpp
@@ -16,7 +16,7 @@ public:
         find_longest_binary_gap();
     }
 
-    int get_longest_binary_gap() const
+    size_t get_longest_binary_gap() const
     {
         return longest_binary_gap;
     }
@@ -24,7 +24,7 @@ public:
 private:
     int number;
     string binary;
-    int longest_binary_gap;
+    size_t longest_binary_gap;
 
     string convert_int_to_binary_string(const int number)
     {
@@ -33,25 +33,25 @@ private:
 
     void find_longest_binary_gap()
     {
-        for (vector<int> index {0, 0, 0}; ranges::find(index, -1) == index.end();) {
-            const int gap_length {find_gap_length(index)};
+        for (vector<size_t> index {0, 0, 0}; ranges::find(index, -1) == index.end();) {
+            const size_t gap_length {find_gap_length(index)};
             longest_binary_gap = {max(gap_length, longest_binary_gap)};
         }
     }
 
-        int find_gap_length(vector<int>& index) const
+        size_t find_gap_length(vector<size_t>& index) const
         {
-            const auto not_found = static_cast<int>(string::npos);
-            index.front() = {static_cast<int>(binary.find('1', index.at(1)))};
+            const auto not_found = string::npos;
+            index.front() = {binary.find('1', index.at(1))};
 
             if (index.front() != not_found) {
-                index.at(1) = {static_cast<int>(binary.find('0', index.front() + 1))};
+                index.at(1) = {binary.find('0', index.front() + 1)};
 
                 if (index.at(1) != not_found) {
-                    index.back() = {static_cast<int>(binary.find('1', index.at(1) + 1))};
+                    index.back() = {binary.find('1', index.at(1) + 1)};
 
                     if (index.back() != not_found) {
-                        const int gap_length {index.back() - index.at(1)};
+                        const size_t gap_length {index.back() - index.at(1)};
                         return gap_length;
                     }
                     else
